VO上传

+++ src/main/java/xmu/crms/vo/CourseDescriptionVO.java
@@ -1,5 +1,7 @@
 package xmu.crms.vo;
 
+import java.math.BigInteger;
+
 /**
  * 
  * @author zyx
@@ -8,7 +10,7 @@ package xmu.crms.vo;
  */
 public class CourseDescriptionVO {
 	
-      private Integer id;
+      private BigInteger id;
       private String name;
       private String description;
       private String teacherName;
@@ -17,7 +19,7 @@ public class CourseDescriptionVO {
 	public CourseDescriptionVO() {
 	}
 
-	public CourseDescriptionVO(Integer id, String name, String description, String teacherName, String teacherEmail) {
+	public CourseDescriptionVO(BigInteger id, String name, String description, String teacherName, String teacherEmail) {
 		this.id = id;
 		this.name = name;
 		this.description = description;
@@ -25,11 +27,11 @@ public class CourseDescriptionVO {
 		this.teacherEmail = teacherEmail;
 	}
 
-	public Integer getId() {
+	public BigInteger getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(BigInteger id) {
 		this.id = id;
 	}
